Change Order Request
Form to describe the requested changes, and to break down the impact of changes on additive and deductive costs in terms of labor, materials, and equipment. Add up the general contractor’s costs, the subcontractor’s costs, and the markup for each, and get final approval before proceeding with changes.